[cgautils] rename utils to cgautils
move cudautils to cgautils folder
