+++ common/cgautils/include/cgautils/signed_integer_utils.hpp
