+++ common/cgautils/include/cgautils/stringutils.hpp
