+++ cudaaligner/benchmarks/singlebatchalignment/CMakeLists.txt
@@ -27,6 +27,6 @@ set(SOURCES
 set(LIBS
     cudaaligner_internal
     cudaaligner
-    utils)
+    cgautils)
 
 cga_add_benchmarks(${PROJECT_NAME} "cudaaligner" "${SOURCES}" "${LIBS}")
