+++ cudaaligner/benchmarks/singlebatchalignment/main.cpp
@@ -12,8 +12,8 @@
 #include <cuda_runtime_api.h>
 #include <random>
 
-#include "utils/genomeutils.hpp"
-#include "cudautils/cudautils.hpp"
+#include "cgautils/genomeutils.hpp"
+#include "cgautils/cudautils.hpp"
 #include "cudaaligner/aligner.hpp"
 #include "aligner_global_ukkonen.hpp"
 #include "aligner_global_myers.hpp"
