+++ cudaaligner/src/aligner_global.hpp
@@ -13,7 +13,7 @@
 #include "cudaaligner/aligner.hpp"
 #include "ukkonen_gpu.cuh"
 #include "device_storage.cuh"
-#include <utils/signed_integer_utils.hpp>
+#include <cgautils/signed_integer_utils.hpp>
 #include <thrust/system/cuda/experimental/pinned_allocator.h>
 
 namespace claragenomics
