+++ cudaaligner/src/aligner_global_myers.cpp
@@ -10,7 +10,7 @@
 
 #include "aligner_global_myers.hpp"
 #include "myers_gpu.cuh"
-#include <utils/mathutils.hpp>
+#include <cgautils/mathutils.hpp>
 #include "batched_device_matrices.cuh"
 
 namespace claragenomics
