+++ cudaaligner/src/alignment_impl.cpp
@@ -9,7 +9,7 @@
 */
 
 #include "alignment_impl.hpp"
-#include <utils/signed_integer_utils.hpp>
+#include <cgautils/signed_integer_utils.hpp>
 
 namespace claragenomics
 {
