+++ cudaaligner/src/myers_gpu.cu
@@ -13,11 +13,11 @@
 #include <climits>
 #include <vector>
 #include <numeric>
-#include <utils/signed_integer_utils.hpp>
-#include <utils/mathutils.hpp>
+#include <cgautils/signed_integer_utils.hpp>
+#include <cgautils/mathutils.hpp>
 #include "device_storage.cuh"
 #include "batched_device_matrices.cuh"
-#include <cudautils/cudautils.hpp>
+#include <cgautils/cudautils.hpp>
 
 namespace claragenomics
 {
