+++ cudaaligner/tests/Test_AlignmentImpl.cpp
@@ -12,7 +12,7 @@
 
 #include "gtest/gtest.h"
 #include "../src/alignment_impl.hpp"
-#include <utils/signed_integer_utils.hpp>
+#include <cgautils/signed_integer_utils.hpp>
 
 namespace claragenomics
 {
