+++ cudapoa/benchmarks/multi-batch/multi_batch.hpp
@@ -12,7 +12,7 @@
 #include <numeric>
 #include "cudapoa/batch.hpp"
 #include "../common/utils.hpp"
-#include "utils/signed_integer_utils.hpp"
+#include "cgautils/signed_integer_utils.hpp"
 
 namespace claragenomics
 {
