+++ cudapoa/benchmarks/single-batch/single_batch.hpp
@@ -10,7 +10,7 @@
 
 #include "cudapoa/batch.hpp"
 #include "../common/utils.hpp"
-#include "utils/signed_integer_utils.hpp"
+#include "cgautils/signed_integer_utils.hpp"
 
 namespace claragenomics
 {
