+++ cudapoa/src/cudapoa_generate_msa.cu
@@ -9,7 +9,7 @@
 */
 
 #include "cudapoa_kernels.cuh"
-#include "cudautils/cudautils.hpp"
+#include "cgautils/cudautils.hpp"
 #include <time.h>
 #include <stdio.h>
