+++ cudapoa/tests/Test_CudapoaNW.cpp
@@ -9,11 +9,11 @@
 */
 
 #include "gtest/gtest.h"
-#include "../src/cudapoa_kernels.cuh"     //runNW, CUDAPOA_*
-#include <cudautils/cudautils.hpp>        //CGA_CU_CHECK_ERR
-#include <utils/stringutils.hpp>          //array_to_string
-#include <utils/signed_integer_utils.hpp> //get_size
-#include "sorted_graph.hpp"               //SortedGraph
+#include "../src/cudapoa_kernels.cuh"        //runNW, CUDAPOA_*
+#include <cgautils/cudautils.hpp>            //CGA_CU_CHECK_ERR
+#include <cgautils/stringutils.hpp>          //array_to_string
+#include <cgautils/signed_integer_utils.hpp> //get_size
+#include "sorted_graph.hpp"                  //SortedGraph
 
 namespace claragenomics
 {
