+++ cudapoa/tests/Test_CudapoaTopSort.cpp
@@ -10,10 +10,10 @@
 
 #include "gtest/gtest.h"
 #include "cudapoa/batch.hpp"
-#include "../src/cudapoa_kernels.cuh"     //runTopSort
-#include <cudautils/cudautils.hpp>        //CGA_CU_CHECK_ERR
-#include <utils/stringutils.hpp>          //array_to_string
-#include <utils/signed_integer_utils.hpp> //get_size
+#include "../src/cudapoa_kernels.cuh"        //runTopSort
+#include <cgautils/cudautils.hpp>            //CGA_CU_CHECK_ERR
+#include <cgautils/stringutils.hpp>          //array_to_string
+#include <cgautils/signed_integer_utils.hpp> //get_size
 
 namespace claragenomics
 {
