+++ cudapoa/tests/basic_graph.hpp
@@ -11,8 +11,8 @@
 #include <string>
 #include <vector>
 #include <stdint.h>
-#include "../src/cudapoa_kernels.cuh"     //CUDAPOA_MAX_NODE_EDGES, CUDAPOA_MAX_NODE_ALIGNMENTS
-#include <utils/signed_integer_utils.hpp> //get_size
+#include "../src/cudapoa_kernels.cuh"        //CUDAPOA_MAX_NODE_EDGES, CUDAPOA_MAX_NODE_ALIGNMENTS
+#include <cgautils/signed_integer_utils.hpp> //get_size
 
 namespace claragenomics
 {
